requirements.txt
Werkzeug==0.16.0
wrapt==1.11.2
tensorflow==2.0.1
six==1.12.0
scipy==1.4.1
PyYAML==5.1.2
pytz==2019.3
python-dateutil==2.8.0
pyparsing==2.4.2
pycparser==2.19
protobuf==3.11.2
pathlib==1.0.1
Pillow==6.2.0
pandas==0.25.1
numpy==1.16.5
opencv-python==4.1.2.30
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
kiwisolver==1.1.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
joblib==0.13.2
itsdangerous==1.1.0
Jinja2==2.10.3
h5py==2.9.0
grpcio==1.16.1
greenlet==0.4.15
google-pasta==0.1.8
gevent==1.4.0
gast==0.2.2
Flask==1.1.1
cycler==0.10.0
cffi==1.12.3
Click==7.0
astor==0.8.0
absl-py==0.8.1
